composeBox, iOS: Fix msg input is not scrolled on typing.
Add `justifyContent` as `flex-end` in `composeText`, which is a wrapper
style of inputs (topic and msg).

Default value is `flex-start`. Which indicates that place components at
the start of the parent and grow in the other direction if required.
With `flex-end`, it means place component at the end of the parent and
grow at the start if required.

Fixes: #3614