webview build: remove dependency on $CWD being a specific value
Make all values explicitly absolute paths. Avoid assuming that $CWD is
any particular value in this script (at least until the auxiliary
scripts are run).

This also fixes a mostly-latent bug, in which `$dest` (if supplied as
a relative path by a human) would be absolutized based on the project
root rather than on $CWD.