presence: strip throttling from `reportPresence` wrapper
The throttling logic currently loses, rather than delaying, an `idle`
which terminates a stream of `active` presence reports. Remove it.

Technically, this alone would fix #3699 as literally reported, since
the `idle` report is now emitted. However, the `active`-emitting
`setInterval` callback is still present and may still fire off events.
(Whether it actually does or not is presumably up to the underlying
OS.)