docs/webview: Write up how we use px and rem (and no em), and why.
We had the key elements of this reasoning, between the existing
version of this doc and the commit messages for 6ef229e42,
ebaffb456, and 3dc7fee83.  But when seeking to explain it
recently I found that not all the dots were connected, plus the
writeups were scattered through those several places.

So, write up the reasoning more fully and in one, editable, place.
Also add its bottom line to our nascent style guide.