composeBox: Set maxHeight of composeBox to 60%.
Now max height is set to composeBox, so when msg input expands it takes
all the space available in the parent and pushes topic input outside.

The workaround of this is to set position of topic input absolute and
adjust position of msg input accordingly with the help of marginTop.

Fixes: #3614