logging: prefer `extras` argument to variable-text messages
When logging information or errors, prefer to use the new `extras`
parameter to pass per-event data, rather than injecting it directly
into the event message string. This should allow Sentry to coalesce
these presumably-related events into single issues, rather than
creating new issues for each event.

Note that we _do_ leave per-event data in the message string if it is
projected to be likely that different values of that data represent
different underlying errors. (This isn't a clear-cut rule, but the
consequences of getting it wrong are also unlikely to be
catastrophic.)