webview build: delete staged files less often
Using the `--delete-excluded` option everywhere causes the auxiliary
scripts to be noncommutative: in particular, the `00-base` script will
delete the preexisting `katex` subdirectory from the staging
directory, even when there was no need.

Remove it from the common invocation, and specify it explicitly only
for the separate staging-to-destination rsync.

----

N.B.: This leaves open the possibility that stale files will remain in
the staging directory, forgotten yet quietly relied upon. This will
need to be addressed -- but addressing it isn't expected to involve
significant changes to the auxiliary build scripts.