welcome: Cut out welcome-help screen.
This was originally added in commits 7082967a6^..9e43aeae0, in
2018-05.  It's pretty bare, and we never came back to really flesh
it out to be more thoroughly helpful; and it doesn't particularly
correspond to any screen seen in the onboarding flows of comparable
apps.  It also sometimes causes confusion in Apple's review process
for releasing updates to the App Store.

Simplify by cutting it out.  This makes the "welcome" screen
redundant, so cut that too and take the user straight to RealmScreen
to enter the URL for their Zulip organization.