nav: Add a hack to determine if RealmScreen allows going "back".
It's now possible to see this as the initial screen, in which case
it should pass `canGoBack: false` to avoid showing the back-arrow
in the app bar; but it's still also possible to navigate here
through the "Add new account" button, in which case we still need
`canGoBack: true`.

This feels like information we really ought to be able to get from
the navigation framework itself, rather than wire up ourselves.  But
I don't see a way to do that; in particular `props.navigation.state`
only has information about the current screen itself, and not about
the history that led to it.

So, here's a hack.

There are a couple of possible directions we could go to make this
less of a hack, or replace it with a non-hack; details in comment.