emoji: Use emoji type to choose either from image or unicode emoji.
Pass `codeToEmojiMap` to `createIconSet`.

createIconSet needs a map where value is emoji and key can be anything
that we pass to refer that emoji. Before we were using emoji name as
key, but it is seen that emoji name can be changed from time to time and
server to server. But code will remain same.

Also messages contain emoji code with them in reactions. So use them instead
of relying on codes hardcoded in the app.

This fixes the issue when emoji names in app are outdated.

Fixes: #3716