users [nfc]: Expand jsdoc and comments on some selectors.
Add various points that I noticed weren't already written down.
Also expand on the meaning of `getUserIsActive`, and give its jsdoc
a summary line.  And we now have a nice issue to point to for the
task of converting to user IDs from emails-as-IDs!

The explanation of what `is_active` means is adapted from the
comment in the server's zerver/models.py (already linked to from
`User`, but that's a couple of indirections away, and this is so
counterintuitive.)