docs/release: Submit early for App Store review.
This saves some latency.  If we had a lot of beta releases we never
send to prod, there might be value in skipping App Store review for
those, but we don't.

UI description is a bit approximate because I don't have it in front
of me.  When I do the next release, I'll be able to see the UI again
and can fill in the exact text of things.