Revert "message list: Modify "Reply" action when in announcement-only narrow."
This reverts commit 674f9a228 (#3811).

That commit fixed only some cases of #3810: it relies on the current
narrow rather than the specific message, and so it misses relevant
messages when we're in a starred narrow, a search, etc.  The right fix
will involve plumbing a different set of data down through these
layers, rather than `isAnnouncementOnly`, so it'll be simpler to do
that fresh than starting from here.