narrow: Make `getNarrowForReply` stop returning stream narrows.
There's no reason we would sometimes want a stream narrow here. This
is just how it was when this function was first written (as
`getNarrowFromMessage`), for narrowing to a message's conversation,
in 34ab1f940 (#333) in 2017.

This case also should be impossible anyway, because message topics
should always be nonempty -- we convert empty to "(no topic)".

Co-authored-by: Chris Bobbe <cbobbe@zulip.com>