narrow: Fix isNarrowValid for group PMs.
This was unintentionally failing to handle group PMs, so that it
unconditionally returned true for them, unlike 1:1 PMs.  Fix that.

Oddly, there was a test checking that this behavior was this way, so
it was noticed before.  It's not clear why it was thought desirable,
though.  The test was added in cefbe6d8b in 2018-05, long after the
logic itself was written in b0c7323d1 in 2017-12, and neither commit
explains the discrepancy.