outbox [nfc]: Replace narrowContains with something more explicit.
This was never quite an accurate fit for what we need here: it gives
the wrong answers for the @-mentions narrow.

More broadly, it's sort of an accident that the inclusion relation on
*narrows* even mostly corresponds to the question we're actually
asking here.  We're really interested in a narrow on one side, and an
individual *message* (specifically an outbox message) on the other.