msglist [nfc]: Factor out "keep WebView at a single page" logic
This code has really quite a different character from the rest of what
the MessageList component is doing, and even more so from what the
rest of the render method is doing.  In fact it isn't really even
about the message list at all: it's about using a WebView to show a
single page, with all navigation handled externally.

In its current location, it puts a wide separation between different
parts of the actual message-list logic, making it harder to read,
understand, or modify.

So, make both this code and the actual message-list code easier to
read, understand, and modify by separating them out from each other.

Also take the opportunity to write some jsdoc explaining what this is
supposed to do.