queueMarkAsRead: Adapt to use Lolex in tests.
Simplify tests significantly by using Jest.

In addition to changes in the tests proper, this also sets the
internal variable `lastSentTime` to an initial value of `-Infinity`
rather than `0`: Lolex's fake clock is initially set to a time of `0`,
which badly confuses the original code.

(An alternative solution would be to bump Lolex's initial time to some
value greater than 0, but this seems more clearly a hack.)