events [nfc]: Reduce cyclomatic complexity of eventToAction.
Replace the explicit list of `actionTypeOfEventType` keys with a check
for presence in `actionTypeOfEventType`.

This is, in a human sense, not really less complex: it's just
switching out a number of identical `case` entries for a table lookup.
Still, as far as ESLint's concerned, it brings the cyclomatic
complexity down from 22 to 13.