lint: Lower function complexity cap.
Prevent any functions from having a cyclomatic complexity of more than
15.

There are currently 4 functions in the codebase with larger cyclomatic
complexity than this:

  * src/message/messageUpdates.js:
    - getMessageTransitionProps()
  * src/session/sessionReducer.js:
    - the default export
  * src/streams/StreamItem.js:
    - StreamItem.render()
  * src/utils/internalLinks.js:
    - getLinkType()

Some of these follow internal patterns meaning that their computed
cyclomatic complexity is misleading; others are actually as
complicated as they are computed to be.

Each has been given a lint-disabling note and an admonishment to
refactor.