search [nfc]: Reduce wrapper/inner distinction to its essential part.
There's one really important thing that the distinction between
these two methods is doing -- really important to notice in order to
understand and correctly modify the code, at any rate.  That's that
the inner method is async, and the outer one is a non-async function
that calls it and then barrels on, ignoring the returned promise.

The two methods also contain different pieces of some of the logic
that relies on `id` and maintains `this.lastIdReceived` and
`state.isFetching`.  But that separation is a cost, not a benefit --
the various spots that set, e.g., `state.isFetching` can only be
understood in combination with each other.

(I think at one point in the history of PR #3623 which led to this
code, an inner/outer separation like this was also used in some
logic for throttling requests.  We ended up leaving that out,
though, and didn't end up determining if this was the best place to
draw a separation for that purpose.  If in the future we do add some
throttling here, we'll be best off in any case starting from code
that's as clear as we can make it, and finding the refactors we need
from there.)

So, reduce the distinction to the one essential thing it's doing, by
pushing everything else inside the inner method.  Now that the outer
method has just one job, it also gives us a natural spot for a
comment about that subtle async point.  And the logic of what's
happening becomes easier to track by being grouped closer together.

This change is NFC because when a JS async function is invoked, its
body is synchronously executed up to the first `await` (or `return`,
or thrown exception) reached.  Everything that previously executed
before the `handleQueryChangeInner` call now executes at the
beginning of that call, and still executes before `handleQueryChange`
itself returns.

With this change, it becomes natural to put the outer function later
in the code, following our usual ordering where callees come first
and then callers; so do that too.