i18n: Offer translations into 8 more languages!
Namely Finnish, Galician, Korean, Lithuanian, Persian, Romanian,
Swedish, and Ukrainian.  Some of these translations were contributed
recently, while others have present in the code for a while... but
we weren't actually offering them for users to choose in our
settings UI.  Oops.  Fix that. :-)

This is part of #3895; apart from the two TODOs noted below, it gets
us back to fully synced with the current state of what's translated.