languages [nfc]: add autogeneration script
Create a script (intended for use immediately after a Transifex pull)
to create automatic imports of all available Transifex message
databases.

This script makes no attempt to name languages in human-readable
fashion (unless the human in question can read ISO-639 codes), nor to
distinguish between languages we should and should not expose for
selection. These remain the responsibility of settings/languages.js.

It is also a known failing of this script (as it is of the system
which it automates) that all messages are loaded at app start time,
rather than only when the relevant language is selected.