i18n [nfc]: Add translation-support-file autogeneration script.
Create a script (intended for use immediately after a Transifex pull)
to create automatic imports of all relevant Transifex message
databases.

It is a known failing of this script's output -- as it is of the
mostly-handwritten files which it replaces -- that all messages for
all languages are loaded at app start time, rather than only once (and
only if) a language is selected for use.