message/webview: Flatten unnecessary nesting in getHtmlPieceDescriptors.
This was just getting flattened out again in renderMessagesAsHtml,
the only place it's being used, so it makes sense not to nest it in
the first place.

Facilitates work later in this series of commits on #3851, to apply
minimal edit sequences to the elements in the message list, instead
of replacing the entire HTML body.

Delete some tests in webViewHandleUpdates-test.js that will be
obsolete following that later work on #3851, rather than rewrite
them for the flattened format.