message/webview [nfc]: Rename renderMessages to getHtmlPieceDescriptors.
Also:

getRenderedMessages -> getShownPieceDescriptorsForNarrow
renderedMessages -> htmlPieceDescriptors

None of this code is directly rendering anything to the screen, so,
reduce that confusion. Where a distinction between "shown" items and
non-shown items is available and effected by the word "rendered"
(i.e., the getRenderedMessages selector), use "shown", instead,
following the example of getShownMessagesForNarrow.