generateContentDiffEvents [nfc]: Add, to soon supersede all-at-once HTML replacement.
To avoid replacing the entire HTML body on every update to the
message list, it's helpful to have a diffing algorithm that can map
changes in the list of shown pieces to specific instructions for DOM
manipulation.

This is our implementation plan:

1) diff previous and current values of the MessageList's
`htmlPieceDescriptors` prop, with help from `backgroundData` and
`narrow`,

2) send these diffs as `WebViewInboundEvent`s to the WebView, and

3) handle those events inside the WebView by adding, replacing, or
deleting DOM nodes.

In this commit, do just 1), and prepare for 2) with commented-out code.

We considered Myers' diffing algorithm from "An O(ND) Difference
Algorithm and Its Variations," but found we could do better. Our
pieces are always in order, by messageId (and within a messageId, by
'time' < 'header' < 'message'*), so we can do it in linear time.

*It's therefore crucial to maintain a consistent ordering within
these pieces, so, comment and test for that.