msglist: Make isFetching be informed by state.session.loading.
When we have no messages to show (i.e. `haveNoMessages` is true) but
we're in the middle of loading a fresh copy of the overall app state
from the server (i.e. `state.session.loading` true), it's more
informative to emphasize the "we're loading messages" rather than
"we have no messages."  Do that.

See further discussion here:
  https://github.com/zulip/zulip-mobile/pull/3822#issuecomment-586541115

The window that this closes will be much smaller following the next
commit, but it's still worth closing. And while this will hide
*some* stale data that users might want to see while waiting for a
refresh, it won't hide much. In particular, showMessagePlaceholders
won't be true unless haveNoMessages is also true, and it's rare to
be looking at a MessageList that has no messages.