lolex: Wrap setTimeout and setInterval to error on overflow.
Lolex follows Node's behavior in (silently!) limiting timer durations
to a maximum of INT32_MAX (2^31 - 1) milliseconds (or approximately
24.86 days), substituting a timeout of 1ms on overflow. Browsers have
subtly different behavior here, as they truncate to 32 bits, rather
than substituting.

React Native's JavaScript engine probably follows browser behavior
rather than Node's, but it hardly matters -- if we issue a timer with
a duration over the limit, even in tests, we're almost certainly not
going to get the expected behavior.

Wrap Lolex's overrides of setTimeout and setInterval to complain if a
timer is ever created with an incompatibly-long timeout duration.