notifications: Accept null device token.
When no registration token is available, Android may report a null
token. This is has been observed to happen at least on an emulator
without Google Play services, and is reported to happen on physical
devices without GP, as well as on devices without Internet access.

Accept null device tokens without comment, effectively returning to
pre-#3888 behavior for them. Adjust the types of `gotPushToken` _et
al._ and the docstring of `pushToken` to reflect this.

Additionally, as issue #3672 is resolved, we no longer need an
explanatory comment pointing to it. Remove the now-confirmed-
unnecessary check for the exact string `"[Object object]"`.