js [nfc]: Use `??` to simplify a few expressions.
Originally part of #3912, in which this was performed manually.
Candidate expressions were identified with a multiline search for
    /==\s*undefined\s*\?/
and manually checked. (`null` was also checked for, but no
simplifiable expressions using it were found.)

This revised commit was generated in part automatically, with a
temporary installation of jscodeshift. The resultant script:

 * correctly excluded one item noted to be incorrect in #3912;
 * arguably correctly excluded another item *not* noted to be
   incorrect;
 * converted one item in a technically-semantically-correct fashion
   that caused Flow to complain and required manual adjustment to fix;
 * had to repeatedly reinvoke Flow via the shell to get type
   information; and
 * was ~140 lines of ugly code (more, if properly cleaned up) that
   doesn't belong anywhere in our repo.