recipient [nfc]: Move filteredRecipientsForPM next to its peers.
This function is very closely related to a few others we have, which
all live in a common file.  That file's name probably isn't the best
it could be, and certainly those functions' names and interfaces
aren't the best they could be... but they're what we have, so start by
just putting things next to each other, and we can go from there.