recipient [nfc]: Split up pm{Ui,Key}RecipientsFromMessage.
These two have the exact same behavior at present... but that's
basically a coincidence.

The "users, except self, unless self-PM" chain of multiple
conditions is desirable for the UI.  But when what we want is a nice
canonical key to use in data structures, it's just unhelpful
complexity and a likely source of bugs -- so it'd be good to migrate
our various data structures over to something simpler, and then
these two will differ.

Before we can attempt any such migration, one thing we need to do is
to separate where we're actually supplying the UI (and therefore
shouldn't change) vs. supplying internal data structures (and
therefore should.)  Do that here.

Also write jsdoc accordingly.