unread [nfc]: Simplify an identity-optimization condition.
It happens to be a fact about removeItemsFromArray that the lengths
of these arrays will be equal just if the arrays are identical.

But if that weren't true, if for example instead of just removing
items we were doing something that could both add and remove items,
then just comparing lengths here would give wrong results,
corrupting the data.

Better to do the simpler thing, that's definitely correct without
having to think about the specifics of what we've done to this
array.

In this version, we do rely on a different fact about
removeItemsFromArray: namely, that if the arrays have *equal
contents*, then they will actually be identical.  But there's a very
important difference, which is that if the assumption fails then
that only defeats an optimization, and doesn't affect the
correctness of our results.