state [nfc]: Eliminate unused `isActive` session state.
`isActive` mirrors and caches React Native's `AppState.currentState`,
and, in turn, the activity state of the app.

However, that underlying state affects the cached value's staleness
period in ways that are nontrivial to reason about. (See parent commit
for a motivating example.) The state in question is also always
globally accessible (whether we like it or not), as it's provided by
React Native itself.

As use of this Redux-cached value has subtle costs and caveats, and as
its only remaining client has just been rewritten to no longer use it,
remove `SessionState.isActive` and all its associated code.

----

Note: this does remove our theoretical ability to use Redux-based
"time-travel debugging" to diagnose issues arising with, or near, uses
of the app's activity state. If we had more extensive logic depending
on it, and if we ever used time-travel debugging in the first place,
this might be worthwhile to preserve.

However, we have never yet had reason to do this. We also probably
never will, even if we begin making extensive use of time-travel
debugging: making *any* decision based on the app's activity state is
necessarily racy (again, see parent commit), and should be done
judiciously.