topic actions [nfc]: Simplify signature of `deleteMessagesForTopic`.
`deleteMessagesForTopic` currently takes both a `Narrow` and a
`Message | Outbox`. However, this is redundant; all the relevant
information can be obtained from either one.

Furthermore, neither is really necessary: we can just ask the caller
to extract the data and hand it to us. Since the caller already knows
that it's operating on a `stream` message, it's actually less trouble
than constructing the `Narrow`.