navigation: Fix back navigation from screens
at present navigations are recorded inside a stack, as the result
whenever user visits a screen will be pushed to stack, when user
presses back, top nav  will be popped. But when 2 screens are inter-
linked, in that case if user tried to navigate sequencially
(mainScreen => a => b => a => b), it becomes hard to navigate back.
He has to come by reverse order to reach back
(b => a => b => a => mainScreen).This looks very hard for use.

So pop untill the first existence of a navigation Screen.

Fixes: #3941