messageActionSheet tests: Push `flags` data down to where used.
The way this file had worked for a while was that this flags object,
with these particular arbitrary message IDs, was used globally in
the file and then these two particular tests used message IDs that
just happened to have specific relationships to those here.

That's no good. :-)  It makes these tests harder to read, by obscuring
their relationship to the data they're actually testing; and it makes
all the other tests harder to read, by raising the possibility that
they might somehow depend on these oddly-specific `flags` values.

Instead, push this data down into the specific tests that use it.
Much clearer.

Also demonstrate how to avoid the Flow error: write the key as an
actual number.  The trick is just to tell ESLint to go away when it
tries to rewrite the key as a (numeric-looking) string. [1]

[1] https://github.com/facebook/flow/issues/380

[chris: link to Flow issue in code and commit message]