redux types: Make zulipVersion string | null.
Keys with value `undefined` don't survive round-trips into JSON, and
the zulipVersion will be persisted in JSON. Use `null`, to maintain
that Account objects always have the same number of keys, and we
don't have to worry about the distinction between an optional
`string` property and a property that's `string | void`.

The same strategy is in use for the ackedPushToken.