redux: Store parsed ZulipVersion in Redux state, instead of string.
Storing the parsed ZulipVersion instance in Redux makes it easier to
work with, so we don't have to think about constructing a new
instance as part of every server feature check.

The one tricky bit is that our Redux state is serialized with
JSON.stringify by redux-persist so that it can be stored in
ZulipAsyncStorage, but the ZulipVersion instance itself is not
serializable. Thankfully, it can be fully represented by its raw
version string, which is serializable.

So, on entry into ZulipAsyncStorage, convert the ZulipVersion
instance into the raw version string with .raw(), and on exit, call
the ZulipVersion constructor on that raw version string. We use
redux-persist's `createTransform`.

We locate a value to apply this transformation to with an
"outside-in" approach, in particular, by the value's extrinsic
property of where it is in the Redux store. That is, we know that
it's at the `zulipVersion` key on all Account objects in
state.accounts.

Instead, we might have used an "inside-out" approach to locate a
value by its *intrinsic* property of being a ZulipVersion instance.

(We don't insist on using a single approach for all future
transforms, but it's useful to compare the two because it's possible
that some kinds of transformations fit one way better than the
other. See #3950, "Use Map rather than object-as-Map in our Redux
state", and #3949, "Use good data structures in Redux state, to
avoid quadratic work", for more things we'll need to transform.)

The "outside-in" approach requires that we write and maintain code
to establish the location of the values to transform. But the
complexity of that code will be no more than the complexity of the
reducer(s) in charge of the same part(s) of the state, so the burden
isn't worrisome. Establishing the location is a two-step process:

1) In src/boot/store, tell the redux-persist transform, created with
createTransform, what key you want to handle. (In this case,
'accounts'.)

2) In the reducer file corresponding to that key, write "save" and
"load" functions that know where to find the value. (They also do
the actual transformation.)

It's helpful in step 2 to consolidate code into the reducer files,
which already have code that's aware of the structure of the part of
state being treated. (Further consolidation like this is #3934.)

If we had instead gone with the "inside-out" approach (i.e.,
locating the value by its being `instanceof ZulipVersion`), which
doesn't depend on knowing the value's location in the Redux state,
we might have used a handy feature of `remotedev-serialize` called
"custom serialization functions" (see
https://github.com/zalmoxisus/remotedev-serialize#passing-custom-serialization-functions).
As noted in #3949, at some point we'll probably use
remotedev-serialize for the particular purpose of transforming
Immutable.js data structures, so it may be convenient to piggy-back
on that for this custom functionality. Greg explains that strategy
at
https://github.com/zulip/zulip-mobile/pull/3952#issuecomment-595497651:

"""
* at save time, values that are `instanceof ZulipVersion` get turned
into objects like `{__serializedType__: 'ZulipVersion', data:
'2.0.0-rc1'}`, before being JSON-serialized

* at load time, after JSON deserialization, values that are objects
with a `__serializedType__` property get transformed in a way
identified by that property's value: in particular if
`__serializedType__: 'ZulipVersion'`, then `data` is fed to `new
ZulipVersion`
"""

This would eliminate the need for custom code that tracks the
location of the values to be transformed. Reasons we don't use this
strategy here are that we don't expect the location of the
ZulipVersion value to change, and that the remotedev-serialize
strategy always requires writing a migration; old data has never
been persisted with that __serializedType__ property.

Fixes: #3951