notifications [nfc]: Break up message-validator function.
... to reduce the detected, if not the effective, cyclomatic
complexity.

This function has a cyclomatic complexity of 18. This is accurate; we
do need at least 18 tests to properly test this function, and breaking
parts of it up into a couple of subroutines doesn't really change that.

It would be simpler (both in general and in this particular case) if
we could use a proper validation library and schema, and somehow
generate either the schema from the associated Flow type or vice
versa.