presence [nfc]: Explain and test handling of timing after stop/start.
Rereading this code today, I spent a few minutes thinking that if we
call `stop` and then `start` again late in an interval, we'd delay the
next heartbeat by another full interval after that.  I.e., if the user
briefly puts the app in the background and comes back, we could send
the next presence heartbeat to the server nearly 2 minutes after the
previous one, potentially causing other clients to see us as offline.

That doesn't happen, but the reason why is subtle, involving a state
where `_active` is false but we still have a live interval counting
down.  Add a comment about this "zombie" state and the whole state
machine, and a test demonstrating the behavior in this case.