msglist: No early return for updateContent event.
The 'content' inbound event, which this series will deprecate, did
part of the duty of the 'fetching' event, such that it didn't matter
if we dropped a 'fetching' event after a 'content' event by
returning early.

The duty in question is to hide the message placeholders. In
updateContent, the HTML was generated with
`nextProps.showMessagePlaceholders`, so the 'fetching' event's
consequence of adding `.hidden` to the message placeholders is
unnecessary.

The edit sequence event should not be responsible for the message
placeholders logic; as designed, it has no awareness of any fetching
state, and it seems more maintainable to separate our concerns this
way (at least without compelling evidence to the contrary).

The 'fetching' event, luckily, already exists, and does just what we
need it to: show and hide the message placeholders, as well as the
top and bottom loading spinners (it's already proven itself to work
for those spinners).

So, make sure it's emitted when appropriate, by not returning early.