msglist [broken]: Send and handle inbound edit sequence events.
Use the edit sequence events from our diffing algorithm to
manipulate the DOM in the WebView.

Still to be handled, later in this series, is proper scrolling:

1) When entering a narrow, whether for the first time or with cached
messages to display, scroll to the "near" message ID if following a
link; otherwise, to the last unread message ID, or the last message.

2) If a new message arrives, and we're near the bottom, auto-scroll
so the new message is in view.

3) Otherwise, preserve the scroll position through any updates to
the message list.

Also to be re-implemented is rewriting image URLs with the API key.