alert word: Add UI to support add, delete alert word
Web app supports "alert word" feature whose primary aim is to
highlight those words and notify to the user when they have occured
in a chat. But user cannot add/remove an alert word using mobile app
as there is no UI to support this feature.
Create an option for this purpose in setting screen. Support
adding and deleting an alert word.

Fixes: #3979