event handlers [nfc]: Add reminder to refactor later.
Replacing our iOS notifications library is already on our to-do list,
and whatever shape that code should be in probably won't look much
like this.

Rather than try to figure out something agnostic, just remind whoever
does that upgrade that it would probably be an opportune time to do a
little more refactoring.