account details: Use 'getUserForEmail' to retrieve user.
Currently, AccountDetailsScreen uses 'getAccountDetailsUserForEmail'
to find the user object for the given email. But this selector is
deprecated - it involves the use of NULL_USER. Replace it with
the new selector 'getUserForEmail' instead.

This can behave differently in the case where the given email
doesn't belong to any known user.  The existing behavior is we'll
show the screen with some dummy data -- like showing the email as
the user's name; trying to use the `realm_uri` as their avatar URL;
and skipping presence information.

That's a pretty broken-looking experience if it can happen; and if
it can happen, it's a bug in the app.  (There's one known case that
can cause it, which is if the user in question changes their email
address.  This is basically part of why #3764 is a bug.)  Instead of
showing some broken-looking bogus data, better to just show nothing.
The simplest version of that is to throw an error; we could do a
softer version of "show nothing", but because this case should be
rare the simple thing is fine for now.

[greg: expanded commit message]