account details: Parse userId in js.js; use new helper.
Moves the user ID parsing logic to 'js.js'. For the actual parsing,
the new helper 'requireNumericAttribute' is used.

'requireNumericAttribute' takes a DOM element and an attribute name
as parameters, and returns the latter's integer parsed value. Throws
if parsing fails. Internally, it calls 'requireAttribute' to get the
string value of the attribute. This will help make parsing integer
attribute values much cleaner.

This approach ensures that the 'crunchy shell, soft center'
principle is followed better.