webview resize, js: Do not scroll when isNearBottom() & difference > 0.
This is a case when user is already near bottom and webview size
increases, one use case is when unread notice pops out.
In such cases, webview content can not further scoll down because it is already at
the end. Thus there is no need of custom handeling in such cases.

This also seems to work fine when keyboard pops in/out.

Fixes: #3301