ios deps: react-native-simple-toast ^0.1.1 -> ^1.0.0
CocoaPods is supported in 0.1.1, but required in 1.0.0. So, now that
we're using CocoaPods, take advantage of the latest version.

At
https://github.com/vonovak/react-native-simple-toast/issues/27#issuecomment-561705243,
the maintainer incorrectly says, "For RN < 0.60, please install the
version before 1.0.0". The correct recommendation would be, "If not
using CocoaPods, please install the version before 1.0.0". The next
comment is correct: "Or you can try installing the native dep
according to the instructions here".

The reason it was failing for that user on 1.0.0 with non-CocoaPods
linking is that a native dependency,
https://github.com/scalessec/Toast, ceases to be bundled with the
app in 1.0.0, and is instead downloaded on running `pod install`,
which of course requires using CocoaPods. (See also the 1.0.0
release notes:
https://github.com/vonovak/react-native-simple-toast/releases/tag/v1.0.0.)

There's no problem using 1.0.0 with RN v0.59.10 as long as you're
using CocoaPods.