UserList: Sort the search result
In the current state, sorting is done in alphabetical order.
If the search term  matches exactly with the username or email,
the matching user may not be listed first.
So prioritize the result of search so that one can find required
result easily
priority will be name startsWith>email startsWith>initial startsWith>
name or email contains with the search term