Language Pick: Implement exact searching of languages
In language pick screen, search result is produced by
checking if the search term is included in any language-string.
But the user expected result may not be in the top of the list.
Since every language strings are smaller in length we can
implement exact searching for a language.

It will list the languages which begins with or whose
secondary name begins with the search term and eliminates
all other languages even if they contain the search term
in the middle.

Fixes: #4014