ios deps: Mock react-native-device-info.
An odd-looking conditional was introduced in 9d9720c; from
experimentation, it looks like this was to prevent some failures
while running our Jest test suite. Some items in NativeModules from
react-native aren't present when Jest is running, so we get errors
if, on startup, we try to run functions that depend on them being
present.

The solution, as we've done for several others, including
`expo-application`, in the parent, is to mock the JS modules that
provide these functions.