ios build: Use CocoaPods 1.9.1.
In 33f4b41c8, we intended to be on 1.9.1, and I did all the work of
that PR using 1.9.1. Greg was using 1.8.4, and when he was fixing up
my branch, he accidentally committed the Podfile.lock with 1.8.4.

But we have no reason not to use this latest version, so, do.