StreamItem [nfc]: Remove `isSelected` prop that is never passed.
I did a search through the whole project for all occurrences of the
string "StreamItem". For each place in the JSX where the component
was created, I did the following checks:

A) By looking at the code, confirmed that we're not passing that
prop ourselves, except for once, in `StreamList`. The value passed
there is based on `StreamList`'s prop `selected`, which is similarly
unused (so, remove that too).

B) By looking at the React Native documentation, confirmed that
FlatList and SectionList (under which StreamItem appears, and which
we don't implement ourselves) don't inject this prop secretly, as
would possible with, e.g., `React.Children` and `cloneElement`.

C) Confirmed (A) and (B) empirically, via the `react-devtools` UI,
by checking that isSelected isn't there.

This was introduced in ec64098a0, for PR #199. As Ray points out
[1], much more recent discussion [2] suggests that we knew it was
dead code, but wanted to preserve it for future functionality. As
Ray summarizes,

"""
a) apparently we used to mimic the webapp's layout more strongly,
with the left and right sidebars being in drawers you could pull out
from either side of the screen, and

b) Boris wanted to reuse this capability for another feature. Those
are both gone, though.
"""

[1]: https://github.com/zulip/zulip-mobile/pull/4046#pullrequestreview-396776517
[2]: https://github.com/zulip/zulip-mobile/pull/3462