unread [nfc]: Handle self-1:1 thread explicitly in getRecipientsIds.
Though the name of this function doesn't yet say so, it's really all
about navigating the `unread` state subtree we maintain from the
server's `unread_msgs` data structure.

Perhaps that's why it hadn't taken account of the possibility of a
self-1:1 appearing -- such a message has to be *sent* by oneself, so
it's a bit of a surprising edge case that it can be unread.  But it
can, if e.g. you send yourself a PM through the API.

Or perhaps the reason is that the code already worked in that case!
This change just makes that more explicit, rather than seeming like
an accident.