styles: Dismantle pipeline for `styles` using legacy Context API.
As foreshadowed in this series, we get to remove the `key={theme}`
hack (along with some boilerplate to get things all working with the
old Context API).

So, also, remove a bit of documentation that explained the hack. The
section was necessary, previously, but it hindered an easy
understanding of the "Pure Component Principle" and our use of it,
which is a main idea of that doc.

Fixes: #1946