deps: Add `immutable` and `zulip/remotedev-serialize@5f9f759a4`.
In an upcoming commit, we will set up a system of "replacer" and
"reviver" transforms so that we can store non-serializable objects,
such as ZulipVersion instances, in Redux, and have a serializable
format be persisted to disk.

This same system will be used for non-serializable objects besides
ZulipVersion instances; see #3949 and #3950, which will be much
easier after this series.

We use our own fork of `remotedev-serialize`, at 5f9f759a4, which
Greg authored atop `zalmoxisus/remotedev-serialize.git` at 0.1.8,
because it fixes an important bug. Also, the project seems
unmaintained, and the maintainer hasn't been active anywhere on
GitHub.

The bug is that data fails to round-trip when it contains a
"__serializedType__" key, which is used by the reviver to identify
and reverse a particular replace transform. Such data, if present,
could trigger hidden functionality.

Also, explain the process of getting the `remotedev-serialize`
libdef set up, in docs/howto/libdefs.md.