typing status: Consume user IDs from events, rather than emails.
We are migrating from emails to user IDs to identify users, and
this is a move in that direction (#3764).

Changes `TypingState` keys to use user IDs rather than emails.
Consequently, changes the consumers of TypingState to handle
this change. This also required replacing ownEmail with ownUserId
in the 'EventTyping' action.

Closes #3930.