theme: Fix latent bug with colors not updating on rerender.
Compute theme-dependent styles in the render method.

As foreshadowed in a previous commit that touched
src/common/Screen.js, fix the remaining latent bugs where the
placement of theme-dependent styles in a class initializer, instead
of the `render` method, meant that these styles don't live-update
when the theme changes, unless the component is unmounted and a new
instance is created to replace it.

The `key={theme}` hack accomplished this replacement, so this bug
hasn't been visible. But that hack will disappear in an upcoming
commit, so this change needs to be made before then.

These bugs were introduced in

51dd1b3b2: LineSeparator.js
a4e0f23b3: OptionButton.js and OptionRow.js
f6ddc2dba: OptionDivider.js