dependencies [nfc] Add js-yaml as an explicit dev-dependency.
In an upcoming commit, we'll add a YAML file to control our set of
available languages. Add this as a dependency so that we can include
it from there.

(This package, and this version thereof, is already used by ESLint.
The absence of a `yarn.lock` change in this commit is due to that, and
is not an error.)