i18n [nfc]: Run the new translation script, and commit its output.
This replaces a number of existing files with autogenerated versions
which have fewer imports. In all cases, these imports were unused.

This also adds a new file, `src/i18n/language_list.js`, which is both
unused and redundant with `src/settings/languages.js`. This will be
amended in an upcoming commit.