ThemeColors [nfc]: Rename to ThemeData.
Best seen in the ThemeContext declaration, below, it's inconsistent
for one name to refer to "colors" and one not to. Furthermore,
"colors" makes it sound like this is the only theme-related data
that exists (which isn't necessarily true), or, worse, that there's
some other theme-related data, besides the colors, that's found
elsewhere, under a different name.

This type encompasses all data about the theme, so, rename it to
ThemeData.

Just `Theme` was considered. `ThemeData` was chosen, not because of
any semantics in Data, but because it's friendlier for newer
contributors who might not be using their "whole-word" setting when
doing a text search through the codebase. With `Theme`, a search
without using that setting would newly contain a lot of noise from,
e.g., ThemeName, ThemeProvider, ThemeContext, etc.

Greg explains some useful searching strategies at
https://github.com/zulip/zulip-mobile/pull/4046#issuecomment-619254798.