account-details-screen: Fix the styling to avoid scrolling
Adding e-mail to the existing details in the profile screen
shifted the components below it further down. This results
in users to scroll down while trying to switch account or
logout.

This commit fixes the issue by decreasing the font size of
all details except the username and also decreasing the
vertical spacing between each details.