redux: Scaffolding for custom replace/revive logic.
This sets up how we'd like to plug in custom replacers and revivers,
but doesn't add any yet.

We can use the '__serializedType__' string freely because we're
using `zulip/remotedev-serialize@5f9f759a4`, which contains a fix
for proper round-tripping of that string, to avoid collisions when
that string occurs as a key representing data, and the unplanned
behavior such collisions would cause.