group-pm: Fix extra spacing at the end of avatars in navbar
There is an extra 16px space after scrolling through the list
of avatars in group-pm navbar, this is caused by adding
marginRight to every component rendered in the Flatlist.

This PR fixes the issue by removing the marginRight and
adding a empty view of width 16 to
ItemSeparatorComponent in the FlatList.