sentry: Add event processor for better fingerprinting of API errors.
Tell the Sentry service that, when bucketing events, it should (also)
consider the error code and HTTP status of any caught `ApiError`s to
be important distinguishing characteristics.

Partially addresses #3864.

Note that, although this is a response to a change made by Sentry to
no longer use error messages as aggregator-hints (at least for custom
error types?), we don't add `.message` to the event fingerprint. The
contents of the `msg` field are mostly intended for humans, and often
contain instance-specific data [1]; when the error message _was_ used
as part of the fingerprint, we had the opposite problem as today --
namely, dozens of different Sentry issues corresponding to the same
bug.

[1] And in the future, it may even be localized: see issue #3692.