api [nfc]: Properly type `apiFetch` and related functions.
`fetch()` accepts many possible extra parameters [1], all of which are
optional. In the absence of an existing type definition for `fetch()`,
a `params` type of `{}` is a reasonably good reflection of that.

However, nowadays we _do_ have type definitions for `fetch()`. Besides
that, in practice, we use only a very small subset of those parameters.
While it was apparently intended to allow `apiFetch` to be called from
arbitrary call sites, that's not functionality we've ever used; since
antiquity [1], we've always opted to go through `apiGet` and its
siblings.

So, encode the tiny slice of `RequestOptions` that we actually use
into `apiFetch`'s parameter type. Anyone who finds they want a larger
subset can expand it as needed.

(And, incidentally, stop exporting apiFetch.)

[1] 999e20a87f2658f2112c0dba10167694c63921eb, 2016-11-11.