sentry: Send up API call and result data on the error event itself.
Sentry error events are often delayed: there may be multiple API
calls to arbitrary routes in between the actual failed call and
the breadcrumb-list being sent.

To avoid any question as to which logged failure resulted in the
current error event, attach the call and response as "extra" data.

See discussion at #4063.