api [nfc]: Clarify distinctions between parameter types.
There are too many things called "params" here, and we're about to add
a third. Let's disambiguate a bit.

`getFetchParams` takes a type that has `FetchParams` in its name and
returns a type that doesn't. Rename it to use the name of the Flow
library type that it actually returns.