ios: Add boilerplate about APIs we don't, in fact, use.
This follows the suggestion of the react-native-unimodules authors:
  https://github.com/unimodules/react-native-unimodules/blob/14f192c1f/README.md

These are needed because (a) some of the code in react-native-unimodules
apparently refers to these APIs, and (b) for Calendars, Contacts, and
Microphone in particular, Apple refuses to allow an app to be uploaded
if it has code referring to these APIs and doesn't have a purpose string.
That means this blocks uploading a release for iOS.

We don't actually ever *invoke* any of that code (barring some very
surprising bug in react-native-unimodules.)  As a result these strings
will never actually be shown to the user.  A better solution would be
to find a way to leave that code out of the build, so that the Apple
robot doesn't mistakenly think we might use it and nag us for these
strings.  But this will do for a minimal fix.