people auto complete: Also send user ID for users having same full name.
Previously, our @-mention autocomplete had a bug, wherein it
failed to correctly mention the specific user if more than one
users had the same full name.

Now, if more than one users have the same full name, we send the
user ID as well to ensure that the correct user is @-mentioned.

The user ID is not sent for every @-mention to ensure the
behaviour is consistent with the web app, i.e., we send it
only in case of ambiguity.

See the `get_mention_syntax` function in `static/js/people.js`
in the webapp source for the origin of this syntax.