i18n types: Fix variance in type of withGetText.
The returned component would have its props types not read-only.
This causes errors if you then try to pass it to `connect`, which
requires its argument's props' types to be read-only -- as in general
the props of a React component should be.

This was invisible because the one place we've been actually using
`withGetText` is `MessageList`... and type-checking there is
defeated by the lack of a libdef for `connectActionSheet`.  It also
wouldn't appear if you pass the result to `connect()(…)` -- it's
only when `connect` gets a `mapStateToProps` callback that the error
would appear.

The root issue here is a Flow bug, that `$Diff` loses the variance
in its arguments:
  https://github.com/facebook/flow/issues/6225
We've actually run into that bug before, and worked around it for
`connect` itself in 0a5a3eec1.  We can do the same thing here.