selectionList [nfc]: Introduce new common selection list component.
This is similiar to language picker, which can be used then multiple
options are available. It is made of `Selector` & `OptionsScreen`.

* `Selector` is the component which shows the label text for the
selection and currently selected value. On clicking it user will
get navigate to `OptionsScreen`.

* `OptionsScreen` shows all the options available, and user can select
on any option by clicking on it.

Our use case for this is to add functionaility to select stream post
policy while creating/editing it. See next commit.