connect types: Give explicit type argument on no-argument `connect`.
Without this change, Flow infers an inappropriately general type for
the resulting components.  In particular, hovering over the call site
of `ComposeMenu` in `ComposeBox.js` shows the type:

    React$Element<ComponentType<empty>>

and if you remove one of the required props, no error appears.

After this change, the type shown on hover there is, appropriately:

    React$Element<
      ComponentType<{|
        +destinationNarrow: Narrow,
        +expanded: boolean,
        +onExpandContract: () => void
      |}>
    >

and removing one of the props from the call produces an error.

I suspect the root of the issue is that with no `mapStateToProps`
argument, there's nothing for the type-checker to hang on to to
constrain what this first type parameter `SP` should be, which is
normally that argument's return type.  The fact that when that
argument is missing it's as if it returned `{}: {||}` is a fact
about the semantics of `connect` that isn't really expressed in the
type we have.  Probably the right fix would be an overload.

This hasn't had much impact because we have only a handful of call
sites of this form, and almost all of the resulting components are
just passed to react-navigation which defeats type-checking anyway.
Issue originally spotted on a newly-added example in development:
  https://github.com/zulip/zulip-mobile/pull/4066#discussion_r423292727