types: Remove all exactness from MessagesState.
MessagesState is exact; until Flow v0.111.0, this is not usefully
compatible with indexer properties. This exactness was added in
8370002d41b7ab0c6f4c8e1d2d8687c008ee9f2e, as part of a mechanical
transformation, before we were aware of this issue.

MessagesState also has `$Exact<Message>` as its indexer value type,
rather than simply `Message`. This is the only use of `$Exact<...>`
directly as the type of a value in our codebase; it was added in
commit 01003e61984c36a5f8a17f0b8edae4c3e7001c47, a professedly
experimental commit, where it is the only such change among several
less interesting `{ ... }` → `{| ... |}` conversions.

Revert both. Add a note about the former and Flow v0.111.0, for
grep's sake.