api: Create helper tryGetFileDownloadUrl.js
Using the new endpoint `/user_uploads/{realm_id_str}/{filename}`
it generates a temporary URL that does not require auth to access.

This allows the user to open the file in the browser without being
logged in to the realm, or to directly download it without
redirects to S3 or needing to send auth headers. If generation of
the url fails, it returns `null`.