deps: Remove `react-navigation-tabs`.
See the previous commit for the reason for the $FlowFixMes.

In 0d2066f41, we followed a warning to use the `react-native-tabs`
package:

```
console.warn node_modules/react-navigation/src/react-navigation.js:180
  TabBarBottom is deprecated. Please use the react-navigation-tabs package
  instead: https://github.com/react-navigation/react-navigation-tabs
```

An important bit of context that's missing from this warning is that
`react-navigation-tabs`'s function `createBottomTabNavigator`,
specifically, is what replaces what was previously done with
`TabBarBottom`.

`react-navigation` also exports another function
`createBottomTabNavigator`. So, is it so important to use the export
from `react-navigation-tabs`?

Turns out it's not; that function in `react-navigation` is an
extremely thin wrapper on the one in `react-navigation-tabs`:

In node_modules/react-navigation/src/react-navigation.js:69:
```
get createBottomTabNavigator() {
  return require('react-navigation-tabs').createBottomTabNavigator;
},
```

So, `react-navigation` depends on `react-navigation-tabs`, so we can
remove the latter as a direct dependency in our project.

The `react-navigation` v2 doc [1] also doesn't say anything about
needing to depend directly on `react-navigation-tabs`. See also
further discussion [2].

It would also have been possible to remove the direct dependency on
`react-native-screens`, but we're planning to use that later; that's
issue #4111.

The same story applies to `createMaterialTopTabNavigator`.

[1]: https://reactnavigation.org/docs/2.x/tab-based-navigation/

[2]: https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/iOS.3A.20r-n-screens.20.2F.20r-n-document-picker/near/877185