MainTabs [nfc]: Prepare for and explain some FlowFixMes.
The next commit's message is quite long, so here's the motivation
for the FlowFixMes.

Importing `createBottomTabNavigator` from `react-navigation` is good
because it means more type checking! Sadly, getting that to really
work in all the ways we want won't be easy.

To sum up, we could get *part* of the way there by making tweaks to
our typed `connect` (and making sure those tweaks didn't break
something or lose type-checking elsewhere!). But there's an
additional hurdle in something that's not quite right in the
`react-navigation` libdef.

The same goes for `createMaterialTopTabNavigator`.

Greg goes into more detail at the PR thread [1].

See also discussion [2].

[1]: https://github.com/zulip/zulip-mobile/pull/4114#issuecomment-634255590

[2]: https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/react-navigation.20types/near/878259